producer/requirements.txt
kafka-python-ng==2.2.3
faker==19.6.2